Return by shared_ptr to avoid copy; add tests to suite.
